+++ src/main/java/com/refinedmods/refinedstorage/mekanism/ChemicalResourceFactory.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage.mekanism.chemical;
+package com.refinedmods.refinedstorage.mekanism;
 
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
@@ -9,12 +9,12 @@ import java.util.Optional;
 import mekanism.api.Action;
 import net.minecraft.world.item.ItemStack;
 
-import static com.refinedmods.refinedstorage.mekanism.chemical.ChemicalResource.ofChemicalStack;
+import static com.refinedmods.refinedstorage.mekanism.ChemicalResource.ofChemicalStack;
 
-public class ChemicalResourceFactory implements ResourceFactory {
+class ChemicalResourceFactory implements ResourceFactory {
     @Override
     public Optional<ResourceAmount> create(final ItemStack stack) {
-        return Optional.ofNullable(stack.getCapability(ChemicalResourceType.ITEM_CAPABILITY))
+        return Optional.ofNullable(stack.getCapability(ChemicalUtil.ITEM_CAPABILITY))
             .map(handler -> handler.extractChemical(Long.MAX_VALUE, Action.SIMULATE))
             .map(chemicalStack -> new ResourceAmount(ofChemicalStack(chemicalStack), chemicalStack.getAmount()));
     }