+++ CHANGELOG.md
@@ -12,3 +12,4 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Chemical Storage Disk
 - Chemical support in the Grid.
 - Chemical support in the Storage Monitor.
+- Chemical support in the Importer.